trivial: Fix some 'Comparison of narrow type with wide type in loop condition'

+++ plugins/synaptics-rmi/fu-synaptics-rmi-common.c
@@ -91,7 +91,7 @@ fu_synaptics_rmi_function_parse(GByteArray *buf,
 		/* set an enable bit for each data source */
 		interrupt_offset = interrupt_count % 8;
 		func->interrupt_mask = 0;
-		for (guint8 i = interrupt_offset;
+		for (guint i = interrupt_offset;
 		     i < (func->interrupt_source_count + interrupt_offset);
 		     i++)
 			func->interrupt_mask |= 1 << i;
